remove constants from routes to config files
