+++ app/get_path.py
@@ -1,0 +1,8 @@
+import os
+
+
+def get_path():
+    os.chdir("files")
+    path = os.getcwd()
+    os.chdir("../")
+    return path
